+++ packages/core/assemblyManager/assemblyManager.ts
@@ -37,12 +37,13 @@ export default function assemblyManagerFactory(
         // hence the union with {name:string}
         const {
           jbrowse: { assemblies },
-          session: { sessionAssemblies = [] } = {},
+          session: { sessionAssemblies = [], temporaryAssemblies = [] } = {},
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
         } = getParent<any>(self)
         return [
           ...assemblies,
           ...sessionAssemblies,
+          ...temporaryAssemblies,
         ] as (AnyConfigurationModel & { name: string })[]
       },
 
@@ -72,14 +73,22 @@ export default function assemblyManagerFactory(
         if (!assemblyName) {
           throw new Error('no assembly name supplied to waitForAssembly')
         }
-        const assembly = self.get(assemblyName)
+        let assembly = self.get(assemblyName)
+        if (!assembly) {
+          try {
+            await when(() => Boolean(self.get(assemblyName)), { timeout: 1000 })
+            assembly = self.get(assemblyName)
+          } catch (e) {
+            // ignore
+          }
+        }
         if (!assembly) {
           return undefined
         }
         await when(
           () =>
-            Boolean(assembly.regions && assembly.refNameAliases) ||
-            !!assembly.error,
+            Boolean(assembly?.regions && assembly.refNameAliases) ||
+            !!assembly?.error,
         )
         if (assembly.error) {
           throw assembly.error