+++ plugins/dotplot-view/src/DotplotRenderer/DotplotRenderer.ts
@@ -225,12 +225,14 @@ export default class DotplotRenderer extends ComparativeRenderer {
           let currX = b1
           let currY = e1
           const cigar = feature.get('cg') || feature.get('CIGAR')
+          const flipInsDel = feature.get('flipInsDel')
 
           if (drawCigar && cigar) {
             const cigarOps = parseCigar(cigar)
 
             ctx.beginPath()
             ctx.moveTo(currX, height - currY)
+
             for (let i = 0; i < cigarOps.length; i += 2) {
               const val = +cigarOps[i]
               const op = cigarOps[i + 1]
@@ -238,14 +240,23 @@ export default class DotplotRenderer extends ComparativeRenderer {
                 currX += (val / hBpPerPx) * strand
                 currY += val / vBpPerPx
               } else if (op === 'D' || op === 'N') {
-                currX += (val / hBpPerPx) * strand
+                if (flipInsDel) {
+                  currY += val / vBpPerPx
+                } else {
+                  currX += (val / hBpPerPx) * strand
+                }
               } else if (op === 'I') {
-                currY += val / vBpPerPx
+                if (flipInsDel) {
+                  currX += (val / hBpPerPx) * strand
+                } else {
+                  currY += val / vBpPerPx
+                }
               }
               currX = clampWithWarnX(currX, b1, b2, feature)
               currY = clampWithWarnY(currY, e1, e2, feature)
               ctx.lineTo(currX, height - currY)
             }
+
             ctx.stroke()
           } else {
             ctx.beginPath()