+++ plugins/linear-genome-view/src/LinearGenomeView/components/util.ts
@@ -51,3 +51,15 @@ export async function fetchResults({
     elt => elt.getId(),
   )
 }
+
+// splits on the last instance of a character
+export function splitLast(str: string, split: string): [string, string] {
+  const lastIndex = str.lastIndexOf(split)
+  if (lastIndex === -1) {
+    return [str, '']
+  } else {
+    const before = str.slice(0, lastIndex)
+    const after = str.slice(lastIndex + 1)
+    return [before, after]
+  }
+}