+++ products/jbrowse-web/src/tests/Alignments.test.tsx
@@ -1,31 +1,19 @@
 import { fireEvent, within } from '@testing-library/react'
-import { LocalFile } from 'generic-filehandle'
 
 // locals
-import { clearCache } from '@jbrowse/core/util/io/RemoteFileWithRangeCache'
-import { clearAdapterCache } from '@jbrowse/core/data_adapters/dataAdapterCache'
 import {
   setup,
   expectCanvasMatch,
-  generateReadBuffer,
   createView,
   hts,
   pc,
+  doBeforeEach,
 } from './util'
 
 setup()
 
 beforeEach(() => {
-  clearCache()
-  clearAdapterCache()
-  // @ts-ignore
-  fetch.resetMocks()
-  // @ts-ignore
-  fetch.mockResponse(
-    generateReadBuffer(
-      url => new LocalFile(require.resolve(`../../test_data/volvox/${url}`)),
-    ),
-  )
+  doBeforeEach()
 })
 
 const delay = { timeout: 20000 }