+++ products/jbrowse-web/src/tests/DrawerWidget.test.tsx
@@ -1,24 +1,12 @@
 import '@testing-library/jest-dom/extend-expect'
 
-import { LocalFile } from 'generic-filehandle'
-import { clearCache } from '@jbrowse/core/util/io/RemoteFileWithRangeCache'
-import { clearAdapterCache } from '@jbrowse/core/data_adapters/dataAdapterCache'
 import { fireEvent, getByRole } from '@testing-library/react'
-import { createView, generateReadBuffer, hts } from './util'
+import { createView, doBeforeEach, hts } from './util'
 
 const delay = { timeout: 15000 }
 
 beforeEach(() => {
-  clearCache()
-  clearAdapterCache()
-  // @ts-ignore
-  fetch.resetMocks()
-  // @ts-ignore
-  fetch.mockResponse(
-    generateReadBuffer(
-      url => new LocalFile(require.resolve(`../../test_data/volvox/${url}`)),
-    ),
-  )
+  doBeforeEach()
 })
 
 test('variant track test - opens feature detail view', async () => {
@@ -39,7 +27,7 @@ test('variant track test - opens feature detail view', async () => {
 }, 20000)
 
 test('widget drawer navigation', async () => {
-  const { view, state, findByTestId, findByText } = createView()
+  const { view, session, findByTestId, findByText } = createView()
   await findByText('Help')
   view.setNewView(0.05, 5000)
   // opens a config editor widget
@@ -62,24 +50,24 @@ test('widget drawer navigation', async () => {
 
   // test minimize and maximize widget drawer
   // @ts-ignore
-  expect(state.session.minimized).toBeFalsy()
+  expect(session.minimized).toBeFalsy()
 
   await findByTestId('drawer-minimize')
   fireEvent.click(await findByTestId('drawer-minimize'))
   // @ts-ignore
-  expect(state.session.minimized).toBeTruthy()
+  expect(session.minimized).toBeTruthy()
 
   fireEvent.click(await findByTestId('drawer-maximize'))
   // @ts-ignore
-  expect(state.session.minimized).toBeFalsy()
+  expect(session.minimized).toBeFalsy()
 
   // test deleting widget from select dropdown using trash icon
   // @ts-ignore
-  expect(state.session.activeWidgets.size).toEqual(2)
+  expect(session.activeWidgets.size).toEqual(2)
   fireEvent.mouseDown(
     getByRole(await findByTestId('widget-drawer-selects'), 'button'),
   )
   fireEvent.click(await findByTestId('ConfigurationEditorWidget-drawer-delete'))
   // @ts-ignore
-  expect(state.session.activeWidgets.size).toEqual(1)
+  expect(session.activeWidgets.size).toEqual(1)
 }, 20000)