+++ products/jbrowse-web/src/tests/TextSearchingImportForm.test.tsx
@@ -1,42 +1,22 @@
-import { waitFor, fireEvent } from '@testing-library/react'
-import { LocalFile } from 'generic-filehandle'
-import { clearCache } from '@jbrowse/core/util/io/RemoteFileWithRangeCache'
-import { clearAdapterCache } from '@jbrowse/core/data_adapters/dataAdapterCache'
+import { screen, waitFor, fireEvent } from '@testing-library/react'
 
 // locals
-import { setup, generateReadBuffer, createView } from './util'
+import { setup, createView, doBeforeEach } from './util'
 
 setup()
 
 beforeEach(() => {
-  clearCache()
-  clearAdapterCache()
-
-  // @ts-ignore
-  fetch.resetMocks()
-  // @ts-ignore
-  fetch.mockResponse(
-    generateReadBuffer(
-      (url: string) =>
-        new LocalFile(require.resolve(`../../test_data/volvox/${url}`)),
-    ),
-  )
+  doBeforeEach()
 })
 
 const delay = { timeout: 10000 }
 
-const total = 30000
-
 async function doSetup(val?: unknown) {
-  const {
-    view,
-    findByText,
-    findByTestId,
-    findAllByText,
-    findByPlaceholderText,
-    getByPlaceholderText,
-  } = createView(val)
+  const args = createView(val)
+  const { view, findByTestId, getByPlaceholderText, findByPlaceholderText } =
+    args
 
+  // clear view takes us to the import form
   view.clearView()
 
   const autocomplete = await findByTestId('autocomplete')
@@ -44,72 +24,62 @@ async function doSetup(val?: unknown) {
     'Search for location',
   )) as HTMLInputElement
 
+  // this will be the input that is obtained after opening the LGV from the import form
+  const getInputValue = () =>
+    (getByPlaceholderText('Search for location') as HTMLInputElement).value
+
   autocomplete.focus()
   input.focus()
 
   return {
-    view,
     autocomplete,
     input,
-    findByText,
-    findAllByText,
-    findByPlaceholderText,
-    getByPlaceholderText,
+    getInputValue,
+    ...args,
   }
 }
 
-test(
-  'test trix on import form (type and hit enter)',
-  async () => {
-    console.warn = jest.fn()
-    const { input, findByText, getByPlaceholderText } = await doSetup()
-    fireEvent.change(input, { target: { value: 'eden.1' } })
-    fireEvent.keyDown(input, { key: 'Enter', code: 'Enter' })
-    fireEvent.click(await findByText('Open'))
-    await waitFor(() => {
-      const n = getByPlaceholderText('Search for location') as HTMLInputElement
-      expect(n.value).toBe('ctgA:1,055..9,005')
-    }, delay)
-  },
-  total,
-)
+test('search eden.1 and hit open', async () => {
+  console.warn = jest.fn()
+  const { input, getInputValue, findByText } = await doSetup()
+  fireEvent.change(input, { target: { value: 'eden.1' } })
+  fireEvent.click(await findByText('Open'))
+  await waitFor(() => expect(getInputValue()).toBe('ctgA:1,055..9,005'), delay)
+}, 30000)
 
-test(
-  'test trix on import form (type and no hit enter)',
-  async () => {
-    console.warn = jest.fn()
-    const { input, findByText, autocomplete, getByPlaceholderText } =
-      await doSetup()
+test('search eden.1 and hit enter', async () => {
+  console.warn = jest.fn()
+  const { input, findByText, autocomplete, getInputValue } = await doSetup()
 
-    fireEvent.mouseDown(input)
-    fireEvent.change(input, { target: { value: 'eden.1' } })
-    fireEvent.keyDown(autocomplete, { key: 'Enter', code: 'Enter' })
+  fireEvent.mouseDown(input)
+  fireEvent.change(input, { target: { value: 'eden.1' } })
+  fireEvent.keyDown(autocomplete, { key: 'Enter', code: 'Enter' })
 
-    fireEvent.click(await findByText('Open'))
-    await waitFor(() => {
-      const n = getByPlaceholderText('Search for location') as HTMLInputElement
-      expect(n.value).toBe('ctgA:1,055..9,005')
-    }, delay)
-  },
-  total,
-)
+  fireEvent.click(await findByText('Open'))
+  await waitFor(() => expect(getInputValue()).toBe('ctgA:1,055..9,005'), delay)
+}, 30000)
 
-test(
-  'lower case refname',
-  async () => {
-    console.warn = jest.fn()
-    const { input, findByText, autocomplete, getByPlaceholderText } =
-      await doSetup()
+test('lower case refname, searching: contigb', async () => {
+  console.warn = jest.fn()
+  const { input, autocomplete, findByText, getInputValue } = await doSetup()
 
-    fireEvent.mouseDown(input)
-    fireEvent.change(input, { target: { value: 'contigb' } })
-    fireEvent.keyDown(autocomplete, { key: 'Enter', code: 'Enter' })
+  fireEvent.mouseDown(input)
+  fireEvent.change(input, { target: { value: 'contigb' } })
+  fireEvent.keyDown(autocomplete, { key: 'Enter', code: 'Enter' })
+  fireEvent.click(await findByText('Open'))
 
-    fireEvent.click(await findByText('Open'))
-    await waitFor(() => {
-      const n = getByPlaceholderText('Search for location') as HTMLInputElement
-      expect(n.value).toBe('ctgB:1..6,079')
-    }, delay)
-  },
-  total,
-)
+  await waitFor(() => expect(getInputValue()).toBe('ctgB:1..6,079'), delay)
+}, 30000)
+
+test('description of gene, searching: kinase', async () => {
+  console.warn = jest.fn()
+  const { input, findByText, getInputValue, autocomplete } = await doSetup()
+
+  fireEvent.mouseDown(input)
+  fireEvent.change(input, { target: { value: 'kinase' } })
+  fireEvent.keyDown(autocomplete, { key: 'Enter', code: 'Enter' })
+
+  fireEvent.click(await screen.findByText('EDEN (protein kinase)', {}, delay))
+  fireEvent.click(await findByText('Open'))
+  await waitFor(() => expect(getInputValue()).toBe('ctgA:1,055..9,005'), delay)
+}, 30000)